requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
click==8.2.1
config-json==0.0.3
Flask==3.1.1
frozenlist==1.7.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
python-telegram-bot==20.3
sniffio==1.3.1
telegram==0.0.1
Werkzeug==3.1.3
yarl==1.20.1
